+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Improved conversation view and styles (@jarekzet, @wujuu)
 - More descriptive error when adding open access to the project second time - User Experience improvement (@danielkryska)
 - Disabled possibility to send an empty resource's review (@goreck888)
+- Clearing logic of adding offers to a project (@kmarszalek)
 
 ### Deprecated